Removed NodeDebugConfigurationProvider until we can provide a cleaner experience
